+++ example/order/index.dart
@@ -9,8 +9,5 @@ part 'model/grid.dart';
 part 'view/board.dart';
 
 main() {
-  var canvas = querySelector('#canvas');
-  var table = new Table.from(new Size.from(4, 4),
-                             new Area.from(canvas.width, canvas.height));
-  new Board(canvas, new TileGrid(table));
+  new Board(querySelector('#canvas'));
 }